Moved carousel to providerInfo
